Updated RideListingPanel
Updated RideListingPanel so that data can actually be added to the table without crashing the program. It uses a factory object to do this. The button has also been implemented but not given a set use other than printing out the name of the person who created the ride offer/request. Will update in the future to open a window that will prompt the user if they want to send the request.