Updates OfferRequestPanel
Updated the OfferRequestPanel to handle invalid input. Handles most invalid cases but will be changed to a more efficience design at a later date. Only thing that needs to be complete is when the submit button is clicked it needs to create the Ride Object and prepare it to be sent to the server.