Finished direct communication between the client and the server
Finished the communication between the client and the server. The client can now receive back an object from the server. Will cause the program to hang if the object is not received properly.

The ride listing panel now receives the listed rides directly from the server. No more "faked" rides.

The next thing to get working will be the log-in functionality.