Updated Server, RideListingPanel, Ride, EmailHandler, and requires login now
Updated email handler to send more descriptive emails to the student that created the ride. Made it so nothing could be used to save or obtain ride information without logging in.  In the ride listing panel, made a helper method to reduce the number of lines that go into creating all the columns. In ride, anything relating to the Button is now set as deprecated. The server has been updated to send emails in a more intelligent way.